Added play, pause, step buttons, and speed control
Sadly i had to compromise and make it impossible to change the speed
while it's running, but it was giving me problems otherwise. It will
probably remain this way, but it's possible i will reapproach this
problem and fix it in the future.

also renamed Main.java to Test.java and cleaned it up a little to look
more like how this project would be used. this included adding a world
folder

Additionally made Main.java in karelz that would run when you double
clicked on the compiled jar, mimicking what Karel J Robot did.