Synced up the play/pause buttons to the methods
Now executing playPause() is the same thing as clicking the play/pause
button if it exists. The simulation also automatically stops when all
the robot threads are stopped. I'm not sure if I like or dislike this
behavior, but I'm keeping it for now as it could be removed by
commenting out one line of code (Window.Java, 879)

additionally made the simulation stop if loading a new world through
"New" or "Load"

I also changed the pop up windows from JFrames to JDialogs so they can
be modals

I added a couple more world constructors, although I'm not sure they
will be used

the "hacky" or "evil" solution I mentioned a couple commits ago is new
present. Window objects now refresh every 100 ms, although at least they
don't when they're not visible. I had to do this because the frame was
getting lots of graphical errors when moving it partially off screen.