ported over room escape code from Karel J Robot to Karel-Z
the previous RoomEscape.java was merged with RoomEscapeRobot.java and
was renamed BlockRoomEscapeRobot.java because it used block walls

the RoomEscapeRobot.java now contains one of my old implementations of
escaping a room in Karel J Robot, which seems to work (most of the
time).

some worlds relevant to the room escape were added

SuperRobots now have the ability to teleport to any valid location

room.kzw was renamed to block-room.kzw and room.kzw is now used for
RoomEscapeRobot.java

NOTE: Window.java shows a lot of changes to the generated code (the
GroupLayout stuff). Nothing was actually changed in that part of the
code.